ASV-1912 - Moved classes
