+++ app/src/main/java/cm/aptoide/pt/home/more/appcoins/EarnAppcListView.kt
@@ -1,9 +1,7 @@
 package cm.aptoide.pt.home.more.appcoins
 
 import cm.aptoide.pt.home.bundles.apps.RewardApp
-import cm.aptoide.pt.home.more.ListAppsClickEvent
-import cm.aptoide.pt.home.more.ListAppsView
-import rx.Observable
+import cm.aptoide.pt.home.more.base.ListAppsView
 
 interface EarnAppcListView : ListAppsView<RewardApp> {
